move modal from NecessityList to PlacaBox
